requirements.txt
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.0
supabase==0.5.8
types-pandas==1.3.3
